# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  /** @minLength 1 */
  title: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  companion: 'companion',
  agent: 'agent',
  system: 'system',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  agentId?: number | null;
  createdAt: string;
}

export interface Memory {
  id: number;
  title: string;
  content: string;
  kind: string;
  importance: number;
  createdAt: string;
  updatedAt: string;
}

export interface MemoryInput {
  /** @minLength 1 */
  title: string;
  content: string;
  kind?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  importance?: number;
}

export interface MemoryUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  kind?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  importance?: number;
}

export type AgentStatus = typeof AgentStatus[keyof typeof AgentStatus];


export const AgentStatus = {
  active: 'active',
  planned: 'planned',
} as const;

export interface Agent {
  id: number;
  name: string;
  role: string;
  description: string;
  status: AgentStatus;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  needs_approval: 'needs_approval',
} as const;

export interface Task {
  id: number;
  title: string;
  objective: string;
  status: TaskStatus;
  /** @nullable */
  agentId?: number | null;
  /** @nullable */
  conversationId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  needs_approval: 'needs_approval',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  objective: string;
  /** @nullable */
  agentId?: number | null;
  /** @nullable */
  conversationId?: number | null;
  status?: TaskInputStatus;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  needs_approval: 'needs_approval',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  objective?: string;
  status?: TaskUpdateStatus;
  /** @nullable */
  agentId?: number | null;
}

export interface Activity {
  id: number;
  type: string;
  summary: string;
  /** @nullable */
  agentId?: number | null;
  /** @nullable */
  taskId?: number | null;
  createdAt: string;
}

export interface CompanionRequest {
  conversationId: number;
  /** @minLength 1 */
  message: string;
}

export type CompanionResponseExecutionState = typeof CompanionResponseExecutionState[keyof typeof CompanionResponseExecutionState];


export const CompanionResponseExecutionState = {
  PROPOSED: 'PROPOSED',
  AUTHORIZED: 'AUTHORIZED',
  EXECUTED: 'EXECUTED',
  OBSERVED: 'OBSERVED',
  EVALUATED: 'EVALUATED',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  RECOVERING: 'RECOVERING',
  ESCALATED: 'ESCALATED',
} as const;

/**
 * @nullable
 */
export type ExecutionTraceJournal = { [key: string]: unknown } | null;

export type ExecutionTraceGraphsItem = { [key: string]: unknown };

export type ExecutionTraceGraphItem = { [key: string]: unknown };

export type ExecutionTraceEventsItem = { [key: string]: unknown };

export type ExecutionTraceActionsItem = { [key: string]: unknown };

export type ExecutionTraceAttemptsItem = { [key: string]: unknown };

export type ExecutionTraceObservationsItem = { [key: string]: unknown };

export type ExecutionTraceEvaluationsItem = { [key: string]: unknown };

export type ExecutionTraceRecoveriesItem = { [key: string]: unknown };

export type ExecutionTraceLessonsItem = { [key: string]: unknown };

export interface ExecutionTrace {
  /** @nullable */
  journal: ExecutionTraceJournal;
  graphs: ExecutionTraceGraphsItem[];
  graph: ExecutionTraceGraphItem[];
  events: ExecutionTraceEventsItem[];
  actions: ExecutionTraceActionsItem[];
  attempts: ExecutionTraceAttemptsItem[];
  observations: ExecutionTraceObservationsItem[];
  evaluations: ExecutionTraceEvaluationsItem[];
  recoveries: ExecutionTraceRecoveriesItem[];
  lessons: ExecutionTraceLessonsItem[];
}

export interface CompanionResponse {
  conversationId: number;
  userMessage: Message;
  companionMessage: Message;
  delegated: boolean;
  agent?: Agent;
  task?: Task;
  reviewed: boolean;
  executionId: string;
  executionState: CompanionResponseExecutionState;
  trace: ExecutionTrace;
}

export interface WorkforceSummary {
  conversationCount: number;
  memoryCount: number;
  openTaskCount: number;
  activityCount: number;
}