# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: number;
  type: string;
  summary: string;
  /** @nullable */
  agentId?: number | null;
  /** @nullable */
  taskId?: number | null;
  createdAt: Date;
}