# lib/api-zod/src/generated/types/agent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentStatus } from './agentStatus';

export interface Agent {
  id: number;
  name: string;
  role: string;
  description: string;
  status: AgentStatus;
}