# lib/api-zod/src/generated/types/companionRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompanionRequest {
  conversationId: number;
  /** @minLength 1 */
  message: string;
}