# lib/api-zod/src/generated/types/executionTrace.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionTraceActionsItem } from './executionTraceActionsItem';
import type { ExecutionTraceAttemptsItem } from './executionTraceAttemptsItem';
import type { ExecutionTraceEvaluationsItem } from './executionTraceEvaluationsItem';
import type { ExecutionTraceEventsItem } from './executionTraceEventsItem';
import type { ExecutionTraceGraphItem } from './executionTraceGraphItem';
import type { ExecutionTraceGraphsItem } from './executionTraceGraphsItem';
import type { ExecutionTraceJournal } from './executionTraceJournal';
import type { ExecutionTraceLessonsItem } from './executionTraceLessonsItem';
import type { ExecutionTraceObservationsItem } from './executionTraceObservationsItem';
import type { ExecutionTraceRecoveriesItem } from './executionTraceRecoveriesItem';

export interface ExecutionTrace {
  /** @nullable */
  journal: ExecutionTraceJournal;
  graphs: ExecutionTraceGraphsItem[];
  graph: ExecutionTraceGraphItem[];
  events: ExecutionTraceEventsItem[];
  actions: ExecutionTraceActionsItem[];
  attempts: ExecutionTraceAttemptsItem[];
  observations: ExecutionTraceObservationsItem[];
  evaluations: ExecutionTraceEvaluationsItem[];
  recoveries: ExecutionTraceRecoveriesItem[];
  lessons: ExecutionTraceLessonsItem[];
}