# lib/api-zod/src/generated/types/memoryUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemoryUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  kind?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  importance?: number;
}