# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MessageRole } from './messageRole';

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  agentId?: number | null;
  createdAt: Date;
}