# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  title: string;
  objective: string;
  status: TaskStatus;
  /** @nullable */
  agentId?: number | null;
  /** @nullable */
  conversationId?: number | null;
  createdAt: Date;
  updatedAt: Date;
}