# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputStatus } from './taskInputStatus';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  objective: string;
  /** @nullable */
  agentId?: number | null;
  /** @nullable */
  conversationId?: number | null;
  status?: TaskInputStatus;
}