# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskUpdateStatus } from './taskUpdateStatus';

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  objective?: string;
  status?: TaskUpdateStatus;
  /** @nullable */
  agentId?: number | null;
}